CONTINUING TO REFACTOR - more symbols moving and naming
